delete comparison code, react-native-compressor
