+++ src/view/com/composer/videos/VideoTranscodeBackdrop.tsx
@@ -1,6 +1,6 @@
-import {clearCache, createVideoThumbnail} from 'react-native-compressor'
 import Animated, {FadeIn} from 'react-native-reanimated'
 import {Image} from 'expo-image'
+import * as VideoThumbnails from 'expo-video-thumbnails'
 import {type QueryClient, useQuery} from '@tanstack/react-query'
 
 import {atoms as a} from '#/alf'
@@ -8,7 +8,6 @@ import {atoms as a} from '#/alf'
 export const RQKEY = 'video-thumbnail'
 
 export function clearThumbnailCache(queryClient: QueryClient) {
-  clearCache().catch(() => {})
   void queryClient.resetQueries({queryKey: [RQKEY]})
 }
 
@@ -16,7 +15,8 @@ export function VideoTranscodeBackdrop({uri}: {uri: string}) {
   const {data: thumbnail} = useQuery({
     queryKey: [RQKEY, uri],
     queryFn: async () => {
-      return await createVideoThumbnail(uri)
+      const result = await VideoThumbnails.getThumbnailAsync(uri)
+      return result.uri
     },
   })
 
@@ -25,7 +25,7 @@ export function VideoTranscodeBackdrop({uri}: {uri: string}) {
       <Animated.View style={a.flex_1} entering={FadeIn}>
         <Image
           style={a.flex_1}
-          source={thumbnail.path}
+          source={thumbnail}
           cachePolicy="none"
           accessibilityIgnoresInvertColors
           blurRadius={15}
